Bug fixes
Fixed bug in main index.html file so that the RT logo now takes
you to the main homepage of whatever the name of the site you
have configured is.
Fixed bug in stbDataPage.pl so an error is displayed if no stb
Grid config file is found (i.e. not been set up yet)