Fixed dusky control bug and ...
Fixed a bug where the last command in a sequence in stbControl.pl would not get through
to an STB on a dusky. A timeout of 1 second was added between sending the last command
and closing the dusky connection.
Increased STB name space from 6 to 8 characters in stbDataPage.pl for better user customisation.