Fixed network control issue
Fixed an issue where control processes might hang when using network
control on a box if the socket connection to the STB failed. Fixed by
adding a 1 second timeout to the initial socket connection. Also changed
the child process for the network control so that it exits with a message
instead of dying