Multiple bug fixes and improvements
Tweaked stbDataPage.pl and stbController.js so that when a user changes the control
type for a STB on the STB Data page, any previous info for that STB and control type
is populated in to the fields. This allows easy switching of control methods for each
STB if required.
eventScheduleControl.pl now manages a debug log file in the files directory called
[scheduledEventDebug.txt] which will allow system admins to identify issues within
their scheduled events. The file is wiped whenever the scheduler is reloaded, an
event is added, removed, disabled, enabled, or the scheduler itself is disabled or
enabled.
Fixed a bug in stbControl.pl which would cause process PIDs to remain listed in the
[pidsRunning] directory when a network control attempt had failed. The method for
dieing when a network socket connection failed was changed to remove the pid file.