Fixed bug in messages.pl
Fixed a bug in messages.pl which related to finding the PID of
the event scheduler process. If other processes listed in the
'ps ax' command contained the digits of the scheduler pid, they
would be shown in the messages section on the page. Adjusted the
grep regex to fix this.