Updated sequences display on page
Changed the css for the sequenceButtons div on the controller page to
limit its height to the same as the controller buttons. Overflow is set
to scroll so if lots of sequences are created, you can scroll down to view
them rather than have them running off the bottom of the page.