Fixed Firefox contenteditable issue
Created a way for the system to cope with issues seen in Firefox around
contenteditable divs. The controller will detect the use of Firefox browsers
and add spaces to buttons that are added to contenteditable divs. This affects
the Sequences, STB Groups, and Events Schedule page.
Also disabled the ability for users to select IR as a control method. This is
because the functionality is not yet developed and may not be.