Changed Dusky control method
Dusky devices are typically controlled via a Moxa device with the port in TCP server mode.
The Moxa devices default to limiting the port to 1 concurrent connection but can support
up to 4 connections. This caused issues when the stbControl.pl script would spawn a
control child process for every STB if there were multiple STBs connected to the same
Dusky. To overcome this, stbControl.pl was changed so that when it receives a command to
control STBs, it puts any STBs that are controlled via Dusky in to a separate hash which is
process differently. A single connection is made to each different dusky that is needed and
then the commands for different STBs on each single Dusky are sent on that single connection.
scripts/videoSwitching.pl was also updated to better handle switches that have IP data of NA.