Fixed clear STB data bug
stbController.js had a bug which meant that clearing all STB data for a box
that had Sky+ Dusky control settings would not work. This was because the
select box for control type was being set to Dusky after the settings were
cleared which meant that the previous settings were reloaded in to the text
fields. This has now been rectified so that STB control type is not changed
when all STB data is cleared.