Lowered dusky connection timeout in stbControl.pl
Changed the connection timeouts for Dusky control in stbControl.pl. Number of retries on a failed connection was reduced from 50 to 5. Dusky connection timeout was reduced from 10 seconds to 2. Dusky connections should not need a long time to connect or many retries. Longer timeouts can cause child processes to hang around for longer than is good.