Updated stbControl function in stbController.js
Changed the way that the stbControl.pl script is requested within stbController.js. The call is now an Ajax GET request which is forced to time out after half a second (500ms). This stops the browser from waiting for a response from the Perl script when STB control actions are performed from the front end. This should improve responsiveness on the interface and lighten the load on the browser when large numbers of separate control requests are triggered.