Add enhanced features for sequences
Add ability to deactivate sequences so that they are not shown on the STB grid list.
Convert sequence data storage to JSON format for better efficiency and control.
Add sequence active state and description to stored data in JSON.
Add feature where sequence description is displayed when the button on the STB grid is hovered over with cursor.
Update sequences page for better css layout and more robust element placing.